+++ CHANGELOG.md
@@ -1,2 +1,9 @@
 
 v0.2.0 - typescript rewrite of v0.1.0 sdk
+
+v0.3.0 - Partial functionality of MockApi
+       - Deprecate:
+        - SDK.Users.fakeCacheWrite()
+        - SDK.Users.indexCacheQ()
+        - SDK.GameBoards.fakeCacheWrite()
+        - SDK.GameBoard.readCacheQ()